requirements.txt
streamlit
supabase
python-dotenv
numpy
Pillow
torch
torchvision
transformers
open_clip_torch

